Chess logic
This is the basic brain of the chess game. It contains a double array
called "board" that holds pointers to "Tiles" which represent a single
square. There are also "Piece" objects, which represent the chess
pieces. A Piece will be found on a Tile, and a Tile will be found
inside of board.

Objects:
Tile
	-Variables
		-loc, a two-element array holding location.
			-For instance, a loc may be [6,3]
			-loc[0] is the row coordinate
			-loc[1] is the column coordinate
		-onMe
			-an array holding all Piece objects on the square.
			-it may have been better to make it something other than an array,
but it's too late.
			-onMe.length is 0 when it is empty, and 1 when there is someone
there.
			-onMe[0] is the Piece on this Tile
Piece
	-Variables
		-type
			-This is a string. Possible values are "pawn, knight, bishop, queen,
king, rook"
		-loc
			-Also the location of the Piece. Just like for Tile
		-color
			-a char, for color. Can be 'b' or 'r' for black or red.
Methods:
	-validateMove( piece, there)
		-Given a particular piece object, and a destination tuple, say,
[4,6], this method will determine whether 		the Piece may move there.
It considers whether the Tile is empty, the type of the Piece, and the
	location of the Piece.

	-move(piece, there)
		-Takes piece, removes it from the tile it's on, and puts it onto the
tile specified by there. It will 		overwrite any piece on there. This
method will work, even if it is not a valid move, and thus
validateMove should be called first.

	-getAvailableMoves(piece)
		-This method will examine piece, and board, and deduce all possible
places the piece could be 		moved. It considers the type of the piece,
the location of the piece, other pieces on the board, and so 		forth.
It returns the places it could go as an array of tuples, for instance [
[3,4] , [5,6] , [2,8] ].